src/requirements.txt
autopep8==1.6.0
bcrypt==3.2.0
bson==0.5.10
cffi==1.15.0
click==8.0.3
colorama==0.4.4
cycler==0.11.0
decorator==5.1.1
distlib==0.3.4
dnspython==2.1.0
email-validator==1.1.3
filelock==3.4.2
Flask==2.0.2
flask-mongoengine==1.0.0
Flask-PyMongo==2.3.0
Flask-WTF==1.0.0
fonttools==4.29.1
idna==3.3
itsdangerous==2.0.1
Jinja2==3.0.3
kiwisolver==1.3.2
MarkupSafe==2.0.1
matplotlib==3.5.1
mongoengine==0.23.1
numpy==1.22.2
packaging==21.3
Pillow==9.0.1
platformdirs==2.4.1
pycodestyle==2.8.0
pycparser==2.21
pymongo==3.12.3
pyparsing==3.0.7
pypng==0.0.21
PyQRCode==1.2.1
python-dateutil==2.8.2
python-dotenv==0.19.2
qrcode==7.3.1
six==1.16.0
toml==0.10.2
validators==0.18.2
virtualenv==20.13.0
Werkzeug==2.0.2
WTForms==3.0.1
